refactor: Migrate script to scripts folder
