+++ src/scripts/update_translation.js
